Fixed datatype issues in Book model and SQL schema causing type casting errors; added seed data to SQL file; confirmed GET all method functions via Swagger

+++ PersonalLibraryApi/Models/Book.cs
@@ -3,7 +3,7 @@
     public class Book
     {
        // title, author, publish date, pages, date acquired, rating, times read
-        public long BookId { get; set; }
+        public int BookId { get; set; }
         public string Title { get; set; }
         public string Author { get; set; }
         public DateTime PublishDate { get; set; }
